metronome update
